requirements.txt
telepot
requests
winshell
tendo
pypiwin32
pillow
pyinstaller
pyaudio